chore: split Pic struct into separate file
